requirements.txt
Faker==37.12.0
apache-airflow-providers-postgres
pyspark==3.5.0
matplotlib